services/security-scanner/requirements.txt

fastapi==0.115.6
uvicorn==0.34.0
langchain-openai==0.2.14
langchain-core==0.3.14
pydantic-settings==2.7.0
python-dotenv==1.0.1
prometheus-client==0.21.1
tiktoken==0.9.0
llmlingua==0.2.2
langfuse==2.57.6
semgrep==1.100.0

